Enum mistralrs::ModelSelected

pub enum ModelSelected {
    Toml {
        file: String,
    },
    Plain {
        model_id: String,
        tokenizer_json: Option<String>,
        arch: NormalLoaderType,
        dtype: ModelDType,
    },
    XLora {
        model_id: Option<String>,
        tokenizer_json: Option<String>,
        xlora_model_id: String,
        order: String,
        tgt_non_granular_index: Option<usize>,
        arch: NormalLoaderType,
        dtype: ModelDType,
    },
    Lora {
        model_id: Option<String>,
        tokenizer_json: Option<String>,
        adapters_model_id: String,
        order: String,
        arch: NormalLoaderType,
        dtype: ModelDType,
    },
    GGUF {
        tok_model_id: Option<String>,
        quantized_model_id: String,
        quantized_filename: String,
    },
    XLoraGGUF {
        tok_model_id: Option<String>,
        quantized_model_id: String,
        quantized_filename: String,
        xlora_model_id: String,
        order: String,
        tgt_non_granular_index: Option<usize>,
    },
    LoraGGUF {
        tok_model_id: Option<String>,
        quantized_model_id: String,
        quantized_filename: String,
        adapters_model_id: String,
        order: String,
    },
    GGML {
        tok_model_id: String,
        tokenizer_json: Option<String>,
        quantized_model_id: String,
        quantized_filename: String,
        gqa: usize,
    },
    XLoraGGML {
        tok_model_id: Option<String>,
        tokenizer_json: Option<String>,
        quantized_model_id: String,
        quantized_filename: String,
        xlora_model_id: String,
        order: String,
        tgt_non_granular_index: Option<usize>,
        gqa: usize,
    },
    LoraGGML {
        tok_model_id: Option<String>,
        tokenizer_json: Option<String>,
        quantized_model_id: String,
        quantized_filename: String,
        adapters_model_id: String,
        order: String,
        gqa: usize,
    },
    VisionPlain {
        model_id: String,
        tokenizer_json: Option<String>,
        arch: VisionLoaderType,
        dtype: ModelDType,
    },
}

Variants

Toml

Select the model from a toml file

Fields

file: String

.toml file containing the selector configuration.

Plain

Select a plain model, without quantization or adapters

Fields

model_id: String

Model ID to load from. This may be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

arch: NormalLoaderType

The architecture of the model.

dtype: ModelDType

Model data type. Defaults to auto.

XLora

Select an X-LoRA architecture

Fields

model_id: Option<String>

Force a base model ID to load from instead of using the ordering file. This may be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

xlora_model_id: String

Model ID to load X-LoRA from. This may be a HF hub repo or a local path.

order: String

Ordering JSON file

tgt_non_granular_index: Option<usize>

Index of completion tokens to generate scalings up until. If this is 1, then there will be one completion token generated before it is cached. This makes the maximum running sequences 1.

arch: NormalLoaderType

The architecture of the model.

dtype: ModelDType

Model data type. Defaults to auto.

Lora

Select a LoRA architecture

Fields

model_id: Option<String>

Force a base model ID to load from instead of using the ordering file. This may be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

adapters_model_id: String

Model ID to load LoRA from. This may be a HF hub repo or a local path.

order: String

Ordering JSON file

arch: NormalLoaderType

The architecture of the model.

dtype: ModelDType

Model data type. Defaults to auto.

GGUF

Select a GGUF model.

Fields

tok_model_id: Option<String>

tok_model_id is the local or remote model ID where you can find a tokenizer_config.json file. If the chat_template is specified, then it will be treated as a path and used over remote files, removing all remote accesses.

quantized_model_id: String

Quantized model ID to find the quantized_filename, only applicable if quantized is set. This may be a HF hub repo or a local path.

quantized_filename: String

Quantized filename, only applicable if quantized is set.

XLoraGGUF

Select a GGUF model with X-LoRA.

Fields

tok_model_id: Option<String>

tok_model_id is the local or remote model ID where you can find a tokenizer_config.json file. If the chat_template is specified, then it will be treated as a path and used over remote files, removing all remote accesses.

quantized_model_id: String

Quantized model ID to find the quantized_filename, only applicable if quantized is set. This may be a HF hub repo or a local path.

quantized_filename: String

Quantized filename, only applicable if quantized is set.

xlora_model_id: String

Model ID to load X-LoRA from. This may be a HF hub repo or a local path.

order: String

Ordering JSON file

tgt_non_granular_index: Option<usize>

Index of completion tokens to generate scalings up until. If this is 1, then there will be one completion token generated before it is cached. This makes the maximum running sequences 1.

LoraGGUF

Select a GGUF model with LoRA.

Fields

tok_model_id: Option<String>

tok_model_id is the local or remote model ID where you can find a tokenizer_config.json file. If the chat_template is specified, then it will be treated as a path and used over remote files, removing all remote accesses.

quantized_model_id: String

Quantized model ID to find the quantized_filename, only applicable if quantized is set. This may be a HF hub repo or a local path.

quantized_filename: String

Quantized filename, only applicable if quantized is set.

adapters_model_id: String

Model ID to load LoRA from. This may be a HF hub repo or a local path.

order: String

Ordering JSON file

GGML

Select a GGML model.

Fields

tok_model_id: String

Model ID to load the tokenizer from. This may be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

quantized_model_id: String

Quantized model ID to find the quantized_filename, only applicable if quantized is set. This may be a HF hub repo or a local path.

quantized_filename: String

Quantized filename, only applicable if quantized is set.

gqa: usize

GQA value

XLoraGGML

Select a GGML model with X-LoRA.

Fields

tok_model_id: Option<String>

Model ID to load the tokenizer from. This may be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

quantized_model_id: String

Quantized model ID to find the quantized_filename, only applicable if quantized is set. This may be a HF hub repo or a local path.

quantized_filename: String

Quantized filename, only applicable if quantized is set.

xlora_model_id: String

Model ID to load X-LoRA from. This may be a HF hub repo or a local path.

order: String

Ordering JSON file

tgt_non_granular_index: Option<usize>

Index of completion tokens to generate scalings up until. If this is 1, then there will be one completion token generated before it is cached. This makes the maximum running sequences 1.

gqa: usize

GQA value

LoraGGML

Select a GGML model with LoRA.

Fields

tok_model_id: Option<String>

Model ID to load the tokenizer from. This may be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

quantized_model_id: String

Quantized model ID to find the quantized_filename, only applicable if quantized is set. This may be a HF hub repo or a local path.

quantized_filename: String

Quantized filename, only applicable if quantized is set.

adapters_model_id: String

Model ID to load LoRA from. This may be a HF hub repo or a local path.

order: String

Ordering JSON file

gqa: usize

GQA value

VisionPlain

Select a vision plain model, without quantization or adapters

Fields

model_id: String

Model ID to load from. This may be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

arch: VisionLoaderType

The architecture of the model.

dtype: ModelDType

Model data type. Defaults to auto.