mistralrs
Function cross_entropy_loss

pub fn cross_entropy_loss(
    inp: &Tensor,
    target: &Tensor,
) -> Result<Tensor, Error>

The cross-entropy loss.

Arguments

[inp]: The input tensor of dimensions N, C where N is the batch size and C the number of categories. This is expected to raw logits.

[target]: The ground truth labels as a tensor of u32 of dimension N.

The resulting tensor is a scalar containing the average value over the batch.