mistralrs::llguidance::api

Enum StopReason

pub enum StopReason {
    NotStopped,
    MaxTokensTotal,
    MaxTokensParser,
    NoExtension,
    NoExtensionBias,
    EndOfSentence,
    InternalError,
    LexerTooComplex,
    ParserTooComplex,
}

Variants

NotStopped

Parser has not emitted stop() yet.

MaxTokensTotal

max_tokens limit on the total number of tokens has been reached.

MaxTokensParser

max_tokens limit on the number of tokens in the top-level parser has been reached. (no longer used)

NoExtension

Top-level parser indicates that no more bytes can be added.

NoExtensionBias

Top-level parser indicates that no more bytes can be added, however it was recognized late.

EndOfSentence

Top-level parser allowed EOS (as it was in an accepting state), and EOS was generated.

InternalError

Something went wrong with creating a nested parser.

LexerTooComplex

The lexer is too complex

ParserTooComplex

The parser is too complex

Implementations

impl StopReason

pub fn is_ok(&self) -> bool