mistralrs::llguidance::derivre

Enum RegexAst

pub enum RegexAst {

    And(Vec<RegexAst>),
    Or(Vec<RegexAst>),
    Concat(Vec<RegexAst>),
    LookAhead(Box<RegexAst>),
    Not(Box<RegexAst>),
    Repeat(Box<RegexAst>, u32, u32),
    MultipleOf(u32, u32),
    EmptyString,
    NoMatch,
    Regex(String),
    SearchRegex(String),
    Literal(String),
    ByteLiteral(Vec<u8>),
    Byte(u8),
    ByteSet(Vec<u32>),
    JsonQuote(Box<RegexAst>, JsonQuoteOptions),
    ExprRef(ExprRef),

}

Variants

And(Vec<RegexAst>)

Intersection of the regexes

Or(Vec<RegexAst>)

Union of the regexes

Concat(Vec<RegexAst>)

Concatenation of the regexes

LookAhead(Box<RegexAst>)

Matches the regex; should be at the end of the main regex. The length of the lookahead can be recovered from the engine.

Not(Box<RegexAst>)

Matches everything the regex doesnโt match. Can lead to invalid utf8.

Repeat(Box<RegexAst>, u32, u32)

Repeat the regex at least min times, at most max times u32::MAX means infinity

MultipleOf(u32, u32)

MultipleOf(d, s) matches if the input, interpreted as decimal ASCII number, is a multiple of d*10^-s. EmptyString is not included.

EmptyString

Matches the empty string. Same as Concat([]).

NoMatch

Matches nothing. Same as Or([]).

Regex(String)

Compile the regex using the regex_syntax crate. This assumes the regex is implicitly anchored. It allows ^$ only at the beginning and end of the regex.

SearchRegex(String)

Compile the regex using the regex_syntax crate, but do not assume itโs anchored. This will add (.*) to the beginning and end of the regex if it doesnโt already have anchors.

Literal(String)

Matches this string only

ByteLiteral(Vec<u8>)

Matches this string of bytes only. Can lead to invalid utf8.

Byte(u8)

Matches this byte only. If byte is not in 0..127, it may lead to invalid utf8.

ByteSet(Vec<u32>)

Matches any byte in the set, expressed as bitset. Can lead to invalid utf8 if the set is not a subset of 0..127

JsonQuote(Box<RegexAst>, JsonQuoteOptions)

Quote the regex as a JSON string. For example, [A-Z\n]+ becomes ([A-Z]|\n)+

ExprRef(ExprRef)

Reference previously built regex

Implementations

impl RegexAst

pub fn contained_in(&self, big: &RegexAst) -> RegexAst

Regex is empty iff self โ big

pub fn get_args(&self) -> &[RegexAst]

pub fn tag(&self) -> &'static str

pub fn write_to_str( &self, dst: &mut String, max_len: usize, exprset: Option<&ExprSet>, )