mistralrs::llguidance::ffi

Function llg_flush_logs

pub extern "C" fn llg_flush_logs(cc: &mut LlgConstraint) -> *const i8

Get the logs from the constraint, since last call to this function. The logs are null-terminated. The logs are kept in the constraint until the next call to this function or until the constraint is freed.