mistralrs

Struct RequestBuilder

pub struct RequestBuilder { /* private fields */ }

A way to add messages with finer control given.

This includes control over:

Logits processors

Constraints

Logprobs

Tools

Sampling

Enable thinking for models that support the configuration

Implementations

impl RequestBuilder

pub fn new() -> Self

pub fn with_web_search_options( self, web_search_options: WebSearchOptions, ) -> Self

pub fn add_message(self, role: TextMessageRole, text: impl ToString) -> Self

Add a message to the request.

For messages with tool calls, use Self::add_message_with_tool_call. For messages with tool outputs, use Self::add_tool_message.

pub fn add_tool_message( self, tool_content: impl ToString, tool_id: impl ToString, ) -> Self

Add a message with the output of a tool call.

pub fn add_message_with_tool_call( self, role: TextMessageRole, text: impl ToString, tool_calls: Vec<ToolCallResponse>, ) -> Self

pub fn add_image_message( self, role: TextMessageRole, text: impl ToString, images: Vec<DynamicImage>, model: &Model, ) -> Result<Self>

pub fn add_audio_message( self, role: TextMessageRole, text: impl ToString, audios: Vec<AudioInput>, model: &Model, ) -> Result<Self>

pub fn add_multimodal_message( self, role: TextMessageRole, text: impl ToString, images: Vec<DynamicImage>, audios: Vec<AudioInput>, model: &Model, ) -> Result<Self>

By convention, all images are added before all audios.

pub fn add_logits_processor( self, processor: Arc<dyn CustomLogitsProcessor>, ) -> Self

pub fn set_adapters(self, adapters: Vec<String>) -> Self

pub fn set_tools(self, tools: Vec<Tool>) -> Self

The default tool choice is auto.

pub fn set_tool_choice(self, tool_choice: ToolChoice) -> Self

pub fn return_logprobs(self, return_logprobs: bool) -> Self

pub fn set_constraint(self, constraint: Constraint) -> Self

pub fn set_sampling(self, params: SamplingParams) -> Self

Set the sampling parameters as given.

pub fn set_deterministic_sampler(self) -> Self

Set the sampling parameters for deterministic generation. This sets up the parameters so that there is:

No temperature, topk, topp, minp

No penalties, stop tokens, or logit bias

No maximum length

pub fn set_sampler_temperature(self, temperature: f64) -> Self

pub fn set_sampler_topk(self, topk: usize) -> Self

pub fn set_sampler_topp(self, topp: f64) -> Self

pub fn set_sampler_minp(self, minp: f64) -> Self

pub fn set_sampler_topn_logprobs(self, top_n_logprobs: usize) -> Self

pub fn set_sampler_frequency_penalty(self, frequency_penalty: f32) -> Self

pub fn set_sampler_presence_penalty(self, presence_penalty: f32) -> Self

pub fn set_sampler_stop_toks(self, stop_toks: StopTokens) -> Self

pub fn set_sampler_max_len(self, max_len: usize) -> Self

pub fn set_sampler_logits_bias(self, logits_bias: HashMap<u32, f32>) -> Self

pub fn set_sampler_n_choices(self, n_choices: usize) -> Self

pub fn set_sampler_dry_params(self, dry_params: DrySamplingParams) -> Self

pub fn enable_thinking(self, enable_thinking: bool) -> Self

pub fn with_truncate_sequence(self, truncate_sequence: bool) -> Self

Truncate prompts that exceed the model’s maximum context length.