mistralrs

Struct TextModelBuilder

pub struct TextModelBuilder { /* private fields */ }

Configure a text model with the various parameters for loading, running, and other inference behaviors.

Implementations

impl TextModelBuilder

pub fn new(model_id: impl ToString) -> Self

A few defaults are applied here:

MoQE ISQ organization

Token source is from the cache (.cache/huggingface/token)

Maximum number of sequences running is 32

Number of sequences to hold in prefix cache is 16.

Automatic device mapping with model defaults according to AutoDeviceMapParams

By default, web searching compatible with the OpenAI web_search_options setting is disabled.

pub fn with_search(self, search_embedding_model: SearchEmbeddingModel) -> Self

Enable searching compatible with the OpenAI web_search_options setting. This loads the selected search embedding reranker (EmbeddingGemma by default).

pub fn with_search_callback(self, callback: Arc<SearchCallback>) -> Self

Override the search function used when web_search_options is enabled.

pub fn with_tool_callback( self, name: impl Into<String>, callback: Arc<ToolCallback>, ) -> Self

Register a callback for a specific tool name.

pub fn with_tool_callback_and_tool( self, name: impl Into<String>, callback: Arc<ToolCallback>, tool: Tool, ) -> Self

Register a callback with an associated Tool definition that will be automatically added to requests when tool callbacks are active.

pub fn with_mcp_client(self, config: McpClientConfig) -> Self

Configure MCP client to connect to external MCP servers and automatically register their tools for use in automatic tool calling.

pub fn with_throughput_logging(self) -> Self

Enable runner throughput logging.

pub fn with_jinja_explicit(self, jinja_explicit: String) -> Self

Explicit JINJA chat template file (.jinja) to be used. If specified, this overrides all other chat templates.

pub fn with_topology(self, topology: Topology) -> Self

Set the model topology for use during loading. If there is an overlap, the topology type is used over the ISQ type.

pub fn with_topology_from_path<P: AsRef<Path>>(self, path: P) -> Result<Self>

Set the model topology from a path. This preserves the path for unload/reload support. If there is an overlap, the topology type is used over the ISQ type.

pub fn with_mixture_qexperts_isq(self) -> Self

Organize ISQ to enable MoQE (Mixture of Quantized Experts, https://arxiv.org/abs/2310.02410)

pub fn with_chat_template(self, chat_template: impl ToString) -> Self

Literal Jinja chat template OR Path (ending in .json) to one.

pub fn with_tokenizer_json(self, tokenizer_json: impl ToString) -> Self

Path to a discrete tokenizer.json file.

pub fn with_loader_type(self, loader_type: NormalLoaderType) -> Self

Manually set the model loader type. Otherwise, it will attempt to automatically determine the loader type.

pub fn with_dtype(self, dtype: ModelDType) -> Self

Load the model in a certain dtype.

pub fn with_force_cpu(self) -> Self

Force usage of the CPU device. Do not use PagedAttention with this.

pub fn with_token_source(self, token_source: TokenSource) -> Self

Source of the Hugging Face token.

pub fn with_hf_revision(self, revision: impl ToString) -> Self

Set the revision to use for a Hugging Face remote model.

pub fn with_isq(self, isq: IsqType) -> Self

Use ISQ of a certain type. If there is an overlap, the topology type is used over the ISQ type.

pub fn with_imatrix(self, path: PathBuf) -> Self

Utilise this imatrix file during ISQ. Incompatible with specifying a calibration file.

pub fn with_calibration_file(self, path: PathBuf) -> Self

Utilise this calibration file to collcet an imatrix. Incompatible with specifying a calibration file.

pub fn with_paged_attn( self, paged_attn_cfg: impl FnOnce() -> Result<PagedAttentionConfig>, ) -> Result<Self>

Enable PagedAttention. Configure PagedAttention with a PagedAttentionConfig object, which can be created with sensible values with a PagedAttentionMetaBuilder.

If PagedAttention is not supported (query with paged_attn_supported), this will do nothing.

pub fn with_max_num_seqs(self, max_num_seqs: usize) -> Self

Set the maximum number of sequences which can be run at once.

pub fn with_no_kv_cache(self) -> Self

Disable KV cache. Trade performance for memory usage.

pub fn with_prefix_cache_n(self, n_seqs: Option<usize>) -> Self

Set the number of sequences to hold in the prefix cache. Set to None to disable the prefix cacher.

pub fn with_logging(self) -> Self

Enable logging.

pub fn with_device_mapping(self, device_mapping: DeviceMapSetting) -> Self

Provide metadata to initialize the device mapper.

pub fn from_uqff(self, path: Vec<PathBuf>) -> Self

๐Deprecated: Use UqffTextModelBuilder to load a UQFF model instead of the generic from_uqff

Path to read a .uqff file from. Other necessary configuration files must be present at this location.

For example, these include:

residual.safetensors

tokenizer.json

config.json

More depending on the model

pub fn write_uqff(self, path: PathBuf) -> Self

Path to write a .uqff file to and serialize the other necessary files.

The parent (part of the path excluding the filename) will determine where any other files serialized are written to.

For example, these include:

residual.safetensors

tokenizer.json

config.json

More depending on the model

pub fn from_hf_cache_pathf(self, hf_cache_path: PathBuf) -> Self

Cache path for Hugging Face models downloaded locally

pub fn with_device(self, device: Device) -> Self

Set the main device to load this model onto. Automatic device mapping will be performed starting with this device.

pub fn with_matformer_config_path(self, path: PathBuf) -> Self

Path to a Matryoshka Transformer configuration CSV file.

pub fn with_matformer_slice_name(self, name: String) -> Self

Name of the slice to use from the Matryoshka Transformer configuration.

pub async fn build(self) -> Result<Model>