Trait mistralrs::Loader

pub trait Loader {
    // Required methods
    fn load_model_from_hf(
        &self,
        revision: Option<String>,
        token_source: TokenSource,
        dtype: Option<DType>,
        device: &Device,
        silent: bool,
        mapper: DeviceMapMetadata,
        in_situ_quant: Option<GgmlDType>
    ) -> Result<Arc<Mutex<dyn Pipeline + Send + Sync>>, Error>;
    fn load_model_from_path(
        &self,
        paths: &Box<dyn ModelPaths>,
        _dtype: Option<DType>,
        device: &Device,
        silent: bool,
        mapper: DeviceMapMetadata,
        in_situ_quant: Option<GgmlDType>
    ) -> Result<Arc<Mutex<dyn Pipeline + Send + Sync>>, Error>;
    fn get_id(&self) -> String;
    fn get_kind(&self) -> ModelKind;
}

The Loader trait abstracts the loading process. The primary entrypoint is the load_model method.

Example

use mistralrs_core::{Loader, TokenSource, DeviceMapMetadata};
use candle_core::Device;

let loader: Box<dyn Loader> = todo!();
let pipeline = loader.load_model_from_hf(
    None,
    TokenSource::CacheToken,
    None,
    &Device::cuda_if_available(0).unwrap(),
    false,
    DeviceMapMetadata::dummy(),
    None,
).unwrap();

Required Methods

fn load_model_from_hf( &self, revision: Option<String>, token_source: TokenSource, dtype: Option<DType>, device: &Device, silent: bool, mapper: DeviceMapMetadata, in_situ_quant: Option<GgmlDType> ) -> Result<Arc<Mutex<dyn Pipeline + Send + Sync>>, Error>

If revision is None, then it defaults to main. If dtype is None, then it defaults to the model default (usually BF16). If model is not found on HF, will attempt to resolve locally.

fn load_model_from_path( &self, paths: &Box<dyn ModelPaths>, _dtype: Option<DType>, device: &Device, silent: bool, mapper: DeviceMapMetadata, in_situ_quant: Option<GgmlDType> ) -> Result<Arc<Mutex<dyn Pipeline + Send + Sync>>, Error>

Load a model from the specified paths. Also initializes DEBUG.

fn get_id(&self) -> String

fn get_kind(&self) -> ModelKind