Trait mistralrs::Pipeline

pub trait Pipeline: Send + Sync {

    // Required methods
    fn forward_inputs(&mut self, inputs: ModelInputs) -> Result<Tensor, Error>;
    fn sample<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        seqs: &'life1 mut [&'life2 mut Sequence],
        logits: Tensor,
        prefix_cacher: &'life3 mut PrefixCacheManager,
        disable_eos_stop: bool,
        rng: Arc<Mutex<Isaac64Rng>>
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait,
             Self: 'async_trait;
    fn device(&self) -> Device;
    fn tokenizer(&self) -> Arc<Tokenizer>;
    fn name(&self) -> String;
    fn get_chat_template(&self) -> Arc<ChatTemplate>;
    fn reset_non_granular_state(&self);
    fn get_metadata(&self) -> &GeneralMetadata;
    fn re_isq_model(&mut self, dtype: GgmlDType) -> Result<(), Error>;
    fn clone_in_cache(
        &mut self,
        seqs: &mut [&mut Sequence],
        modify_draft_cache: bool
    );
    fn clone_out_cache(
        &mut self,
        seqs: &mut [&mut Sequence],
        modify_draft_cache: bool
    );
    fn set_none_cache(
        &mut self,
        reset_non_granular: bool,
        modify_draft_cache: bool
    );
    fn cache(&self) -> &Cache;
    fn activate_adapters(
        &mut self,
        adapters: Vec<String>
    ) -> Result<usize, Error>;

    // Provided methods
    fn step<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 mut self,
        input_seqs: &'life1 mut [&'life2 mut Sequence],
        is_prompt: bool,
        prefix_cacher: &'life3 mut PrefixCacheManager,
        disable_eos_stop: bool,
        rng: Arc<Mutex<Isaac64Rng>>,
        pre_op: CacheInstruction,
        post_op: CacheInstruction
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait,
             Self: 'async_trait { ... }
    fn tokenize_prompt(&self, prompt: &str) -> Result<Vec<u32>, Error> { ... }
    fn apply_chat_template(
        &self,
        messages: Vec<IndexMap<String, String>>,
        add_generation_prompt: bool
    ) -> Result<String, Error> { ... }

}

Required Methods

fn forward_inputs(&mut self, inputs: ModelInputs) -> Result<Tensor, Error>

fn sample<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 self, seqs: &'life1 mut [&'life2 mut Sequence], logits: Tensor, prefix_cacher: &'life3 mut PrefixCacheManager, disable_eos_stop: bool, rng: Arc<Mutex<Isaac64Rng>> ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait, Self: 'async_trait,

fn device(&self) -> Device

fn tokenizer(&self) -> Arc<Tokenizer>

fn name(&self) -> String

fn get_chat_template(&self) -> Arc<ChatTemplate>

fn reset_non_granular_state(&self)

fn get_metadata(&self) -> &GeneralMetadata

fn re_isq_model(&mut self, dtype: GgmlDType) -> Result<(), Error>

fn clone_in_cache( &mut self, seqs: &mut [&mut Sequence], modify_draft_cache: bool )

Clone the cache FROM the sequences’ cache TO the model cache. Only called for completion seqs. It is not a guarantee that this will be called for each completion step.

fn clone_out_cache( &mut self, seqs: &mut [&mut Sequence], modify_draft_cache: bool )

Clone the cache FROM the model cache TO the sequences. Called for prompt and completion seqs. It is not a guarantee that this will be called for each step.

fn set_none_cache(&mut self, reset_non_granular: bool, modify_draft_cache: bool)

Set the model cache to all None. Only called for prompt seqs. It is not a guarantee that this will be called for each prompt step. This may also reset the non granular state if applicable.

fn cache(&self) -> &Cache

fn activate_adapters(&mut self, adapters: Vec<String>) -> Result<usize, Error>

Returns the number of activated adapters.

Provided Methods

fn step<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 mut self, input_seqs: &'life1 mut [&'life2 mut Sequence], is_prompt: bool, prefix_cacher: &'life3 mut PrefixCacheManager, disable_eos_stop: bool, rng: Arc<Mutex<Isaac64Rng>>, pre_op: CacheInstruction, post_op: CacheInstruction ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait, Self: 'async_trait,

This does forward pass of model followed by run.

fn tokenize_prompt(&self, prompt: &str) -> Result<Vec<u32>, Error>

fn apply_chat_template( &self, messages: Vec<IndexMap<String, String>>, add_generation_prompt: bool ) -> Result<String, Error>