mistralrs::v0_4_api

Struct VisionModelBuilder

pub struct VisionModelBuilder { /* private fields */ }

Configure a vision model with the various parameters for loading, running, and other inference behaviors.

Implementations

impl VisionModelBuilder

pub fn new(model_id: impl ToString) -> Self

A few defaults are applied here:

Token source is from the cache (.cache/huggingface/token)

Maximum number of sequences running is 32

Automatic device mapping with model defaults according to AutoDeviceMapParams

By default, web searching compatible with the OpenAI web_search_options setting is disabled.

pub fn with_search(self, search_bert_model: BertEmbeddingModel) -> Self

Enable searching compatible with the OpenAI web_search_options setting. This uses the BERT model specified or the default.

pub fn with_throughput_logging(self) -> Self

Enable runner throughput logging.

pub fn with_jinja_explicit(self, jinja_explicit: String) -> Self

Explicit JINJA chat template file (.jinja) to be used. If specified, this overrides all other chat templates.

pub fn with_prompt_chunksize(self, prompt_chunksize: NonZeroUsize) -> Self

Set the prompt batchsize to use for inference.

pub fn with_topology(self, topology: Topology) -> Self

Set the model topology for use during loading. If there is an overlap, the topology type is used over the ISQ type.

pub fn with_chat_template(self, chat_template: impl ToString) -> Self

Literal Jinja chat template OR Path (ending in .json) to one.

pub fn with_tokenizer_json(self, tokenizer_json: impl ToString) -> Self

Path to a discrete tokenizer.json file.

pub fn with_loader_type(self, loader_type: VisionLoaderType) -> Self

Manually set the model loader type. Otherwise, it will attempt to automatically determine the loader type.

pub fn with_dtype(self, dtype: ModelDType) -> Self

Load the model in a certain dtype.

pub fn with_force_cpu(self) -> Self

Force usage of the CPU device. Do not use PagedAttention with this.

pub fn with_token_source(self, token_source: TokenSource) -> Self

Source of the Hugging Face token.

pub fn with_hf_revision(self, revision: impl ToString) -> Self

Set the revision to use for a Hugging Face remote model.

pub fn with_isq(self, isq: IsqType) -> Self

Use ISQ of a certain type. If there is an overlap, the topology type is used over the ISQ type.

pub fn with_calibration_file(self, path: PathBuf) -> Self

Utilise this calibration_file file during ISQ

pub fn with_paged_attn( self, paged_attn_cfg: impl FnOnce() -> Result<PagedAttentionConfig>, ) -> Result<Self>

Enable PagedAttention. Configure PagedAttention with a PagedAttentionConfig object, which can be created with sensible values with a [PagedAttentionMetaBuilder].

If PagedAttention is not supported (query with paged_attn_supported), this will do nothing.

pub fn with_max_num_seqs(self, max_num_seqs: usize) -> Self

Set the maximum number of sequences which can be run at once.

pub fn with_logging(self) -> Self

Enable logging.

pub fn with_device_mapping(self, device_mapping: DeviceMapSetting) -> Self

Provide metadata to initialize the device mapper.

pub fn from_uqff(self, path: Vec<PathBuf>) -> Self

Path to read a UQFF file from.

pub fn from_max_edge(self, max_edge: u32) -> Self

Automatically resize and pad images to this maximum edge length. Aspect ratio is preserved. This is only supported on the Qwen2-VL and Idefics 2 models. Others handle this internally.

pub fn write_uqff(self, path: PathBuf) -> Self

Path to write a UQFF file to.

The parent (part of the path excluding the filename) will determine where any other files generated are written to. These can be used to load UQFF models standalone, and may include:

residual.safetensors

tokenizer.json

config.json

And others

pub fn from_hf_cache_pathf(self, hf_cache_path: PathBuf) -> Self

Cache path for Hugging Face models downloaded locally

pub fn with_device(self, device: Device) -> Self

Set the main device to load this model onto. Automatic device mapping will be performed starting with this device.

pub async fn build(self) -> Result<Model>