mistralrs::llguidance

Struct TokenParser

pub struct TokenParser {
    pub token_env: Arc<dyn TokenizerEnv + Sync>,
    pub parser: Parser,
    pub compute_mask_start_time: Instant,
    pub last_bias_time: Duration,
    pub inference_caps: InferenceCapabilities,
    pub logger: Logger,
    pub limits: ParserLimits,
    pub bias_computer: Arc<dyn BiasComputer>,
    pub dbg_grammar: String,
    /* private fields */
}

Fields
token_env: Arc<dyn TokenizerEnv + Sync>parser: Parsercompute_mask_start_time: Instantlast_bias_time: Durationinference_caps: InferenceCapabilitieslogger: Loggerlimits: ParserLimitsbias_computer: Arc<dyn BiasComputer>dbg_grammar: String
Implementations

impl TokenParser

pub fn grammar_warnings(&mut self) -> Vec<String>

pub fn get_capture(&self, name: &str) -> Option<&[u8]>

pub fn deep_clone(&self) -> TokenParser

pub fn stop_reason(&self) -> StopReason

pub fn stopped(&self) -> bool

pub fn is_fresh(&self) -> bool

pub fn parser_stats(&self) -> &ParserStats

pub fn last_step_stats(&self) -> &ParserStats

pub fn max_step_stats(&self) -> &ParserStats

pub fn num_tokens(&self) -> usize

pub fn final_bytes(&self) -> &[u8]

pub fn is_accepting(&mut self) -> bool

pub fn bytes_since(&self, idx: usize) -> &[u8]

pub fn start_without_prompt(&mut self)

pub fn process_prompt(&mut self, prompt: Vec<u32>) -> Vec<u32>

pub fn augment_err(&self, e: impl Display) -> String

pub fn dump_state(&self) -> String

pub fn error_message(&self) -> Option<String>

pub fn validate_token(&mut self, token: u32) -> Result<bool, Error>

pub fn reset(&mut self) -> Result<(), Error>

pub fn rollback(&mut self, n_tokens: usize) -> Result<(), Error>

pub fn validate_tokens_raw(&mut self, tokens: &[u32]) -> Result<usize, Error>

Returns how many of the passed tokens can be accepted by the parser. It does not tokenize forced bytes, so will accept non-canonical tokenizations. If called with more than one token, it may ignore max_tokens constraints.

pub fn compute_mask(&mut self) -> Result<SimpleVob, Error>

pub fn force_bytes(&mut self) -> Vec<u8>

pub fn temperature(&self) -> Option<f32>

pub fn consume_token(&mut self, token: u32) -> Result<usize, Error>

Extend the current state of the parser with given token. Returns number of tokens to backtrack if any.

pub fn check_stop(&mut self) -> Result<bool, Error>

Check whether the current parser state forces the sequence to stop. If so, puts the parser in stop state and returns true. Otherwise, returns false. This generally should be called after consume_token().

pub fn compute_ff_tokens(&mut self) -> Vec<u32>

Check if there are any tokens to fast-forward, forced by the current parser state.

pub fn consume_ff_tokens(&mut self) -> Result<Vec<u32>, Error>

Compute and then consume fast-forward tokens.